fix: user RolesForSelect type
