+++ src/middlewares/checkauth.ts
@@ -24,5 +24,23 @@ const checkAuth = (role: Roles) => async (
     next(templateErrors.UNAUTHORIZED('Unauthorized'));
   }
 };
+const checkToken = async (
+  req: RequestWithUserRole,
+  res: Response,
+  next: NextFunction,
+) => {
+  try {
+    const { token } = req;
+    if (token) {
+      const decoded = await verifyToken(token) as Decode;
+      req.user = decoded;
+      next();
+    } else {
+      throw templateErrors.UNAUTHORIZED('Unauthorized');
+    }
+  } catch (error) {
+    next(templateErrors.UNAUTHORIZED('Unauthorized'));
+  }
+};
 
-export default checkAuth;
+export { checkAuth, checkToken };
